+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.0 beta 4
+* Create a default config file if it does not exist
 * Set colors on tooltips too in the default theme
 
 # 1.0 beta 3